Update spaceTimeLags.R
I removed arrange from the main dplyr pipe and added a check at the top of the function. Arranging just by datetime was resulting in a messed up ordering of participant IDs even though the df was just grouped by participant ID. In any case, the data should already be organized in this fashion, so I thought it made more sense to add a check/correction at the top of the function.